refactor: group related files into packages
