+++ src/main/java/Main.java
@@ -1,0 +1,5 @@
+public class Main {
+    public static void main(String[] args) {
+        // start gui here
+    }
+}
