+++ src/main/java/gui/RPanel.java
@@ -1,3 +1,5 @@
+package gui;
+
 import java.awt.Dimension;
 import java.awt.Graphics;
 import java.awt.Graphics2D;
