+++ src/main/java/gui/RTextField.java
@@ -1,3 +1,5 @@
+package gui;
+
 import java.awt.Graphics;
 import java.awt.Graphics2D;
 import java.awt.RenderingHints;
